@@ -3,40 +3,45 @@ import { RouterModule, Routes } from '@angular/router';
 import { NgModule } from '@angular/core';
 import { TopicListingComponent } from './topic-listing.component';
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import {
+  provideHttpClient,
+  withInterceptorsFromDi,
+} from '@angular/common/http';
 import { LetDirective } from '@ngrx/component';
 
 const routes: Routes = [
   {
     path: '',
-    component: TopicListingComponent
+    component: TopicListingComponent,
   },
   {
     path: 'architecture',
-    loadChildren: () => import('../01-architecture/architecture/architecture.module').then(m => m.ArchitectureModule)
+    loadChildren: () =>
+      import('../01-architecture/architecture/architecture.module').then(
+        (m) => m.ArchitectureModule
+      ),
   },
 
   {
     path: 'component',
-    loadChildren: () => import('../02-advanced-component/advanced-component.module').then(m => m.AdvancedComponentModule)
+    loadChildren: () =>
+      import('../02-advanced-component/advanced-component.module').then(
+        (m) => m.AdvancedComponentModule
+      ),
   },
 
   {
     path: 'routing',
-    loadChildren: () => import('../08-advanced-routing/advanced-routing.module').then(m => m.AdvancedRoutingModule)
+    loadChildren: () =>
+      import('../08-advanced-routing/advanced-routing.module').then(
+        (m) => m.AdvancedRoutingModule
+      ),
   },
-
 ];
 
 @NgModule({
-  declarations: [
-    TopicListingComponent,
-  ],
-  imports: [
-    RouterModule.forChild(routes),
-    CommonModule,
-    LetDirective,
-    HttpClientModule
-  ],
+  declarations: [TopicListingComponent],
+  imports: [RouterModule.forChild(routes), CommonModule, LetDirective],
+  providers: [provideHttpClient(withInterceptorsFromDi())],
 })
-export class TopicListingModule { }
+export class TopicListingModule {}
